% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.R
\docType{package}
\name{cma}
\alias{cma}
\alias{cma-package}
\title{cma: COVID mortality Ascertainment}
\description{
cma is a reproducible R compendium for analysing impact of COVID-19 vaccination
to date.
}
\author{
\strong{Maintainer}: OJ Watson \email{o.watson15@imperial.ac.uk}

}
