% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squire_utils.R
\name{seroprev_df}
\alias{seroprev_df}
\title{Extract PCR prevalence and seroprevalence from squire model fit}
\usage{
seroprev_df(res)
}
\arguments{
\item{res}{Output of [[squire::pmcmc]]}
}
\description{
Extract PCR prevalence and seroprevalence from squire model fit
}
